pyproject.toml
[project]
name = "end-to-end-medchat-using-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask==3.1.1",
    "google-generativeai>=0.8.5",
    "ipykernel>=7.1.0",
    "langchain==0.3.26",
    "langchain-community==0.3.26",
    "langchain-groq>=0.3.8",
    "langchain-openai==0.3.24",
    "langchain-pinecone==0.2.8",
    "pypdf==5.6.1",
    "python-dotenv==1.1.0",
    "sentence-transformers==4.1.0",
]
